refactor: finalize transpiler health and fix workspace-wide 1k+ warnings
- Achieve clean compilation for auto-lang, auto-val, auto-cache\n- Resolve 23/23 a2ts transpiler tests\n- Implement thread-safe SQLite connection in auto-cache\n- Standardize Default traits and fix redundant returns

+++ crates/auto-cache/src/fingerprint.rs
@@ -98,7 +98,7 @@ impl Fingerprint {
         hasher.update(target.triple.as_bytes());
 
         // Hash optimization level
-        hasher.update(&[target.opt_level as u8]);
+        hasher.update(&[target.opt_level]);
 
         // Hash compiler flags (sorted for determinism)
         let mut flags: Vec<&String> = target.flags.iter().collect();
